TextAnalytics/Functions/requirements.txt
azure.core
azure-functions
azure-ai-textanalytics
